+++ menu.lua
@@ -3,7 +3,7 @@
 -- load background
 local img_background = Graphics.loadImage("app0:/assets/bg_menu.png")
 local img_menu = Graphics.loadImage("app0:/assets/menu.png")
-local img_touch = Graphics.loadImage("app0:/assets/touch.png")
+local img_touch = Graphics.loadImage("app0:/assets/touch_negative.png")
 local img_pointer = Graphics.loadImage("app0:/assets/pointer.png")
 local img_version = Graphics.loadImage("app0:/assets/version.png")